requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
groq==0.5.0
httpx==0.27.2
python-dotenv==1.0.1
pytest==7.4.0
beautifulsoup4==4.14.3
pdfminer.six==20260107
pytesseract==0.3.13
python-docx==1.2.0
soupsieve==2.8.3
edge-tts==7.2.7
